fix(bump): 2.0.0
